third_party/WebKit/LayoutTests/svg/dom/SVGViewSpec-invalid-ref-crash.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="container"><svg xmlns="http://www.w3.org/2000/svg" width="200" height="200"/></div>
<script>
    description("Confirm that null is returned when an SVGViewSpec is used after its corresponding element has been removed.");

    successfullyParsed = false;

    testRunner.waitUntilDone;
    window.jsTestIsAsync = true;

    var svgView = document.getElementById("container").childNodes[0].currentView;
    document.getElementById("container").removeChild(document.getElementById("container").childNodes[0]);

    asyncGC(function() {
        shouldBeNonNull("svgView.preserveAspectRatio;");
        shouldBeNonNull("svgView.transform;");
        shouldBeNull("svgView.viewTarget;");
        shouldNotThrow("svgView.viewBox;");
        shouldBeEqualToString("svgView.preserveAspectRatioString", "xMidYMid meet");
        shouldBeEqualToString("svgView.transformString", "");
        shouldBeEqualToString("svgView.viewBoxString", "0 0 0 0");
        finishJSTest();
    });

    successfullyParsed = true;

</script>
</body>
</html>
